When adding issue occurrences, mark them additionally as 'added@'
For example if change 4711, patch set 2 comes with occurrences
"somewhere", "body", and "subject", and patch set 3 has occurrences
"somewhere", "body", and "footer" for the same issue, then
additionally, the occurrence "added@footer" is adjoined for patch
set 3.

Thereby, rules can be formulated that fire if and only if an issue
association in some part has been just added to the change.

Change-Id: I8bedba4fbd91f188c4fdf553941273cb3e52389b