Configure branches for ITS integration on init
Add an init step that allows to configure the branches for which the
issue tracker system should be enabled/enforced. Admins can specify a
branch name, a ref pattern or a regular expression. By default
'refs/heads/*' is proposed which excludes the refs/meta/config branch
which is not used for product development and hence normally commits
to it don't need to contain any associations to issues. Still if
refs/meta/config should be included the ref pattern can easily be
changed to 'ref/*'.

The init step only supports the configuration of a single value for
branch. If multiple values are already configured in the
'project.config' file of the All-Projects project the init step
refuses to change the configuration as it doesn't want to overwrite
other values. In this case the init step prints out a message saying
that the branch configuration has to be done manually in the
'project.config' file of the All-Projects project.

Change-Id: Ic815b069dfa600be071f6f27d440cb043e63507b
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>