Update patch set 3
Patch Set 3:

> Like the idea to move to Buck, however this change actually breaks
 > all the other its-* plugins.

Ok, after buckification of its-jira i see the confusion and why we were talking past each other: parent-pom construct wasn't used in its-bugzilla. And that was first its-base depended plugin that was buckified. In its-jira and its-rtc parent-pom was used and this change intact broke Maven build for these plugins. Fixed in [1] and [2].

* [1] https://gerrit-review.googlesource.com/58672
* [2] https://gerrit-review.googlesource.com/58674

Patch-set: 3