Update patch set 6
Patch Set 6: Verified+1

| Would this include the ability to publish the maven artifact as well?

Not yet, but this would be trivial to add. Gitiles project, that is also Bucklets driven, does it already. But again, this is only needed when you would want to build its-base and friends outside of the Gerrit tree, in standalone mode.This is currently not supported, but we may add this mode later.

Patch-set: 6
Label: Verified=+1