Use pluginName directly instead of facade's name
The recent commits that ease setting up the same kind of Its twice
(e.g.: connecting two different Bugzilla instances at the same time to
only a single Gerrit site), blurred the difference between a plugin's
name and the name of the corresponding Its. Hence, we drop this
distinction, and use the plugin's name to identify the Its.

Change-Id: I8ba52f1ccbf38e32357aa3d5077a7d4b119b2de9