Remove plugin name from build configuration
Removing a specified name from the build configuration allows the Gerrit
plugin loader to dynamically set the plugin name to the same as the file
name (minus the extension). This makes it a little more flexible for the
end user.  In cases where the end users want to change the plugin name
they only need to rename the plugin file instead of having to rebuild
the plugin.  Added some documentation for a use case where this would be
very handy.

Change-Id: Ib685a8023ce2c1ce77b5e5a555e424db35fe892a