Fix branch limitation feature in its plugins
The logic to determine if branch is enabled was checking a refPattern
(refs/heads/*) against a branch name (master).  In order to do that
the code would need to extract all branches from the refPattern to
determine if any of the branches matched with the branch that the
change occured in.  This didn't exist so the branch limiting feature
never worked.  This change modifies that logic to only match on
the branch name pattern instead because that's the only information
we get from the stream event.

The initialization code did not support multiple branch settings
in project.config but that is something that is supported now.
Refactor the initialization code to allow for multiple branch
settings.

Updated documentation to match the above changes.

Change-Id: Id98c8bdb11a4184ebdf3a23ef7b4483fb08ee940