Make work with Gerrit core stable-2.10 branch.
The stable-2.10 branch for this plugin was forked from master
which failed to build against Gerrit core stable-2.10.  This
change is a fix to to make this plugin build and work against
Gerrit stable-2.10

The scenarios were manually verified:
1. Buck built the plugin and ran all tests without failure
2. Maven built and ran all tests without failure
3. Built a dependent its-plugin against its-base and verified that
it work as expected.  I used the its-storyboard plugin to test.

Change-Id: Ie492e9021bd61500fb2e2868f0951c0564bf82bd