Change branch limiting feature to enable all branches by default
The default setting for an its-base plugin is to enable "refs/heads/*"
branches.  This default setting will actually disable actions on all
branches due to issue 3099.  Also pointed out in the issue is that
if this config option is not set at all then all branches are
enabled.  This change removes this setting thereby enabling its
actions on all branches.  I think this is the proper default.  If users
want to restrict actions on certain branches then they will have to set
the config.

Bug: issue 3099
Change-Id: I779923320aa4164b11e18445c73a9ca86732364e