Update patch set 3
Patch Set 3:

The change is small and makes sense: my only concern is that we are chaning existing functionality and potentially breaking it for some users.

Can we introduce a configuration parameter to say if we want to use the subject or the whole commit message for it?

Patch-set: 3
Reviewer: Gerrit User 1006192 <1006192@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Verified=0