Update patch set 3
Patch Set 3:

In what way would it break for people? This change would find issue ids in more commits than before, you think that might be a problem for some? Performing issue state transitions or something like that on issues it is not supposed to do so?

Whether adding a configuration option is possible or not is a question you can probably answer better than me, I am not a java developer and not very familiar with the gerrit APIs. But from what I remember about it it probably wouldn't be too hard.

Patch-set: 3