Update patch set 2
Patch Set 2: Code-Review-1

I am all for the move! Thanks.

However, its-base currently has many open changes, and they'd all need
to be redone if we merge this one. Mhmmm. Not sure.
Since this is a quite disruptive change ... could we hold back the
package move of its-base until the other changes have been dealt with?

-----------------------------------------------

I see that moving its-base classes into

  com.googlesource.gerrit.plugins.its

makes sense. But that makes its-base special by populating the main
its package directly. Also, for other Its plugins that are not based on
its-base, it would be strange to have stuff directly in
com.googlesource.gerrit.plugins.its .


By analogy with its-phabricator and its-storyboard, I'd instead prefer
to go with

  com.googlesource.gerrit.plugins.its.base

package.

(Marking CR-1 not because I do not like the change, but just to mark
that the discussion is ongoing)

Patch-set: 2
Label: Code-Review=-1