Lower complexity of individual tests for isEnabled
The isEnabled tests for events used a helper method to abstract the
test setting. While we strive for abstraction in general, we try to
avoid it unneeded abstraction for tests, if it is only saves a few
lines of code. For tests, we rather have them explicit, even if it
means a bit of duplication. This "being explicit" helps to quicker
understand what a test is doing, if a test is failing.

Change-Id: Icdf848998ebdce329ecf5ad6b1808b45eb249b73