Implement standalone Buck driven build
Allow to build the plugin outside of Gerrit tree. Add rules to install
and deploy the final artifacts: plugin, javadoc and sources to Maven
local or remote repositories. This would allow the its-* plugins that
depend on its-base plugin to easily be built outside of the Gerit tree
and without cloning and building its-base plugin first.

Change-Id: Ic415310a61123bb27abde7b45550aaf64b5d556d