Update patch set 1
Patch Set 1:

While I understand your reticence and agree there is a burden keeping up to date, the main goal of this change is to be able to include this plugin in a CI chain. 

While I understand your reticence and agree there is a burden keeping up to date, the main goal of this change is to be able to include this plugin in a CI chain. 

Right now, after eliminating the Maven build (for good!), there is no way for us to produce repeatable, automated builds for this plugin. 
I mean, we use to have Jenkins jobs that used Maven to build plugins but after the dismiss of Maven in most of the plugins (including our 
internal ones), we converted our jobs to buck builds and we use standalone mode to avoid having to checkout the gerrit source code to compile a plugin. The problem is, even if we accept to check it out  every time, which commit/tag do we checkout if we do not have a file (i.e. lib/gerrit/BUCK) that specify what is the gerrit version the plugin depends on?

Ideally, we should work in the near future to improve the standalone mode in order to overcome its pitfalls, but for the moment, I think it's still useful as it is.

Patch-set: 1
Reviewer: Gerrit User 1019328 <1019328@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Verified=0