Update patch set 1
Patch Set 1:

To be honest, that does not convince me to add the manual updating and
all the errors that it'll bring.

The plugins are already in CI systems:

* http://builds.quelltextlich.at/gerrit/nightly/master/latest/index.html#its-base.jar (sources are at http://git.quelltextlich.at/?p=gerrit/gerrit-builder;a=summary )
* https://gerrit-ci.gerritforge.com/view/Plugins-master/job/plugin-its-base-master/ (sourcer are at https://gerrit.googlesource.com/gerrit-ci-scripts )

> there is no way for us to produce repeatable, automated builds for this plugin.
> [...]
> we use standalone mode to avoid having to checkout the gerrit source code [...]

Those two statements do not go together for me.
Standalone mode does not help you to get repeatable builds,
as the API evolves. "2.12-SNAPSHOT" is no repeatable API to build
against. For repeatable builds, you at least need the commit hash for
the API you build against. And then we're arriving at cloning the
gerrit source tree anyways.

So if you really want repeatable builds, you need to clone the gerrit source
tree anyways.

> [...] which commit/tag do we [have to] checkout [...]?

A plugin's HEAD of the `master` branch is expected to build against gerrit's HEAD of the `master` branch.
A plugin's HEAD of the `stable-2.11` branch is expected to build against gerrit's HEAD of the `stable-2.11` branch.
A plugin's HEAD of the `stable-2.10` branch is expected to build against gerrit's HEAD of the `stable-2.10` branch.

Patch-set: 1