Update patch set 1
Patch Set 1:

> I definitely agree with Christian: this isn't a standalone project
 > as it very tied to its associated Gerrit version.

Huh? Every plugin, no matter for what project, is tied to the project it's trying to extend. Yet, normally, it's not necessary to set-up and build the core project to just build a plugin.

I had discussion with Christian in Berlin, during last Gerrit hackathon on that topic, and we actually agree, that standalone build for Gerrit plugin makes a lot of sense. His concern is  (only) to keep the dependencies in sync with gerrit core.

Fortunately, latest Buck version (there are pending changes in Gerrit and Bucklets projects to update Buck) cansimplify that job. For one it allows you to attach (undocumented yet) maven_coord data to java_library() rule and for another it borrowed the query language from Blaze/Bazel build toolchains. With that in place, it should be possible (haven't tried that) to delegate Maven coordinates from our custom maven_jar() invocation to underlying java_library() rule and query it with buck query. The result can be used to sync Maven dependencies between plugin-standalone and gerrit-core Buck build implementations.

Patch-set: 1