Update patch set 3
Abandoned

It seems that there is agreement from the its-base maintainers that
standalone build is currently maintenance overhead. Hence, aborting.

But just to make sure that the work that went into this change has not
been in vain, I added a link to this change in the plugin's BUCK
file [1]. So once standalone got rid of the maintenance overhead, we
can easily find and restore this change.

[1] https://gerrit-review.googlesource.com/71045

Patch-set: 3
Status: abandoned