Buck: Adapt test rules to new acceptance-framework artifact
In I4cadf6616d acceptance-framework was exposed as new library
and Ie9e63de622 where //gerrit-acceptance-tests:lib with all its
transitive dependencies was added to plugin-api was partially
reverted.

Adapt test rules to add GERRIT_TESTS dependency. Note, that this
variable must be used, to suport in tree and standalone Buck build
modes, that will be added in follow-up change.

Change-Id: I77b4249c7fee1036cd38bc5e422c00f393a94adb