Update patch set 5
Patch Set 5:

> Isn't this more or less the same as [1] which was rejected by
 > Christian and Luca?
 > 
 > [1] https://gerrit-review.googlesource.com/#/c/70896/

Not really. What was rejected is the very first patch set of: [1], where substantial parts of the Gerrit core artifacts were unnecessary re-defined in standalone build mode. Follow-up patch sets addressed all comments. Moreover this change incorporates the most recent changes made in Gerrit core, with the split of plugin-api  and acceptance-framework. Before the split and before the refactoring in Gerrit core the standalone unit tests implementation in [1] was broken, because the plugin-api shaded JAR erroneously included bouncy castle signatures in manifest. See the discussion in this thread: [2]. In addition, this change can package and install/deploy the jar, sources and javadoc in local/central Maven repositories. As the matter of fact, in tree build toolchain doesn't support that.

* [2] https://groups.google.com/d/topic/repo-discuss/QEY_eXj-5jU/discussion

Patch-set: 5