Update patch set 5
Patch Set 5:

> (1 comment)
 > 
 > I have to say that I like the idea of building its-base and
 > publishing it to a public maven repository :-)

Yes  ;-) Actually, this is a second reason why standalone build mode is required (while the first is simplification). A new its-foo plugin that may not use Buck driven build depends on its-base. There is currently no way for such a plugin, to consume the final artifact that is located in buck-out/gen/plugins/its-base/... as result of in tree build process. Say, your github-plugin that is Maven driven, would like to consume its_base.jar. How would you do this? With this change you would just add another Maven artifact as a dependency, done. Note, that in (old) Maven driven Gerrit plugin build it was trivial to install/deploy to Maven local and central repositories.

 > 
 > The standalone build however depends on buckets so you *have to
 > clone* another repo anyway, how can you convince me that cloning
 > bucklets is less work than cloning Gerrit? ;-)

This is in fact very good question and there is open issue upstream: [1] to allow/simplify reusing of Buck rule across different projects/repositories. For now all you can do is to clone Bucklets separately (not to compare with cloning and building Gerrit, of course). Another approach would be to add bucklets as submodule to its-base repository. This is what gitiles and gerrit-oauth-provider plugin: [2] are currently doing. With this in place, you would just say:

  $ git clone --recursive https://gerrit.googlesource.com/plugins/its-base
  $ cd its-base && buck build mvn_install

And of course, this would also solve another issue, that you would always reference the correct version of bucklets we are compatible with.

* [1] https://github.com/facebook/buck/issues/116
* [2] https://github.com/davido/gerrit-oauth-provider

Patch-set: 5
Reviewer: Gerrit User 1011323 <1011323@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Verified=0