Update patch set 1
Patch Set 1:

> @Chirstian, why don't you rebase your change on top of Claudio's?

Because Claudio's change is not doing the "simple fix" that you called for either.
The simple fix for

  event.change.project

that comes with least danger would be

  event.change.get().project

. But that's not used in Claudio's change. Instead

  event.getProjectNameKey().get()

is used, which unneededly creates the Project.NameKey instance only
to destroy it again after extracting the `.project` again. That's backwards.

Sure, we could ask Claudio to adapt his change to be the "simple fix"
that you have in mind and avoid even the tiniest refactorings and have
him update to

  event.change.get().project

.

But asking him to spend time on something that we intend to throw away, feels
counter-productive. Because at the end of the day, `.getProjectNameKey()` is
the right thing. The callees can easily work with Project.NameKey objects. At
the end, they need it, so it's even better if they directly operate on
Project.NameKeys.

And to be honest ... "refactoring" is quite a big word here ;-)

Patch-set: 1