Mock CommitReceiveEvent to be master-compatible
Gerrit master has changed the way CommitReceiveEvent gets created.
However, it is not the scope of our test to assert Gerrit's internal
behavior, as it may change across different versions.

By using a mocked event, we can concentrate on what we care and test in
our scenario, and be more Gerrit-independent in the future.

Fix the associated tests on master and stable-2.14

Bug: Issue 6106
Change-Id: I28dd3c4b9c0b635ff42ba0a87a4c0b13791b3588