Add support for closure template (soy)
Breaking change: Variables that had a dash have been renamed to
exclude the dash and instead capitalise the word that used to be after
the dash.

The renaming of variables to remove the dash in them is because
soy does not support variables with a dash in it.

Anyways velocity 2.0 also dropped support for variables with dash too.

Bug: Issue 6528
Bug: Issue 6411
Change-Id: I847d1697e72314d4d30cfa0006c02a7bdc65c53b