Bazel: Depend on last released plugin API version
Normally, when developing on master branch, plugins depend on plugin
API snapshot version to be always up to date with core gerrit.

This approach unnecessary complicates the plugin managements, as it
requires to branch stable branches whenever a new release is cut of,
even though the plugins may be still compatible with older gerrit
version.

Replace this approach to build the plugn against latest reelased gerrit
API version. We could develop this approach further and create different
API versions in bazlets, with release number in their names, so that the
plugins can declare what is the oldest version they can depend on, like:

gerrit_api_2.14.bzl
gerrit_api_2.15.bzl

and so on. We don't have this for now, and Bazel is only supported since
2.14, so switch to consuming the API from the last stable plugin API
release.

Change-Id: I03a023f5362c128f4dccc8261bb344265f31ed95