Update patch set 2
Patch Set 2:

> > If this works but breaks the tests, can this still be merged to allow users to use this with notesdb?
> 
> I think we should stop asking to merge changes when they are not ready. Given changes in gerrit-review are open for non-owners to upload new patch sets, when in a hurry the right thing to do is to make the change ready, IMHO.

Agreed: if tests are broken is because they are either wrong (if the code works and the interface hasn't changed, the test should still work) or the code is wrong. I believe here we are talking about the first case (tests are going too much into the implementation details) and we should fix them before merging this change. Keeping wrong tests in the code-base is the start of the "broken windows" syndrome :-(

Patch-set: 2
Reviewer: Gerrit User 1006192 <1006192@173816e5-2b9a-37c3-8a2e-48639d4f1153>