Update patch set 6
Patch Set 6:

A stable-2.15 branch should be created for this :)

I will backport the soy change once a 2.14 branch is created as i want to get that deployed to wmf when we upgrade. But that needs a 2.14 branch.

please.

Patch-set: 6
Reviewer: Gerrit User 1029953 <1029953@173816e5-2b9a-37c3-8a2e-48639d4f1153>