Update patch set 1
Patch Set 1:

We are currently blocked on upgrading zuul, which the maintainer of that for wmf dosen't have time to do.

So i am figuring out how to unblock it for them. It seems simple just to revert this (not merge it here) but for use at wmf.

Patch-set: 1
Reviewer: Gerrit User 1029953 <1029953@173816e5-2b9a-37c3-8a2e-48639d4f1153>