Disable firing RefUpdatedEvent on change creation
RefUpdatedEvent should not be fired on creating a new change.
Previously in gerrit version stable-2.12,  RefUpdatedEvent was not
fired for "refs/changes/*". This changed in Ie9a132 and duplicated the
actions defined for the previous "ref-updated" event. This behaviour
is not desirable in ITS-plugins e.g. its-rtc.

Fixing the its-base plugin to adopt the new behaviour of gerrit's
"ref-updated" event prevents duplication of code required to fix the
same issue in individual plugins.

Change-Id: I622029e26a6d96ef4389e10aaffce5a594a3f1bd