RFC:Support connection to multiple its-servers in its-base
Pass information required to connect to multiple servers in child
its-plugins. Current server info is passed with every action performed
in any child its-plugins.

The information about servers is not accessed at any place in its-base.
The interface `ItsServer` facilitates the tunnelling of information to
and from its-base. The child its-plugins can implement both single or
multiple servers as per the requirements.

Change-Id: I4d3a1710caa0817faf610f191c4d798a9a43d5ff