Support connection to multiple its-servers
So far, its-base derived plugins were able to connect only to an unique
server which configuration was defined in 'gerrit.config'. In enterprise
environments, it is not unusual to have multiple instances of the same ITS
serving different clients. In case of ITS-* plugins, the only way to
achieve the integration of different clients with Gerrit was to deploy
different copies of the its plugin with different names.

This change is an attempt to bring a more flexible way of supporting
multiple instances of the same ITS. Given the information about the
project name is handled by its-base and not passed directly to the
derivative plugins, a new interface, `ItsFacadeFactory`, facilitates the
tunnelling of information to and from its-base.

The child its-plugins can implement both single or multiple servers as
per requirements. In the first case, they only need to bind the
interface to the 'UniqueItsServer' implementation an no further code
change is required. In the multiple servers case, they need to implement
the interface `ItsFacade` in a way that it captures the project name
passed by its-base to the its-child plugin, and the
latter will send back the `ItsFacade` object to its-base.

Change-Id: I4d3a1710caa0817faf610f191c4d798a9a43d5ff