Support connection to multiple its-servers
So far, its-base derived plugins were able to connect only to an unique
server which configuration was defined in 'gerrit.config'. In enterprise
environments is not unusual to have multiple instances of the same ITS
serving different clients. In this case, the only option to achieve the
integration with Gerrit was to deploy different copies of the its plugin
with different names.

This change is an attempt to bring a more flexible way of supporting
multiple instances of the same ITS. Given the information about the
project name is handled by its-base and not passed directly to the
derivative plugins, a new interface, `ItsServer`, facilitates the
tunnelling of information to and from its-base.

The child its-plugins can implement both single or multiple servers as
per requirements. In the first case, they only need to bind the
interface to the 'UniqueItsServer' implementation an no further code
change is required. In the multiple servers case, they need to implement
the interface to pass the project name to the its-child plugin, and the
latter will send back the respective server info to its-base so that it
can be passed as a parameter to any action on the ITS system.

Change-Id: I4d3a1710caa0817faf610f191c4d798a9a43d5ff