Simplify the event enabling code by allowing RefEvents only
Move the event type check in ActionController and pass only RefEvents
to the code enabling events in ItsConfig.

Modify the PropertyExtractor class to comply with event type changes.

This refactoring is needed work in order to ease the handling of
multiple servers which is done in a follow up change.

Change-Id: I038493bbea0bc91742c1dea64f2c88cfa1b7c41e