Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1: Code-Review-2
> > 
> > Can you please cherry pick to master? This seems to me much more than a simple and urgent bug fix
> 
> I've said this before.  I don't think we should be restricting plugin stable branches to bugfixes only.  The only reason we have these branches is to maintain compatibility/buildability against the corresponding core Gerrit version.  Plugins should be allowed to develop new features without forcing users to also upgrade to a later version of core Gerrit (unless of course the feature depends on a core functionality).

There is nothing that forbids plugins on their master version to be compatible with multiple versions of Gerrit. See for instance the websessions-flatfile :-)

Why don't we then cherry-pick this to master and propose a change that makes it work with Gerrit 2.14?

The development of a product on multiple diverging branches becomes a big headache when trying to merge them together. Additionally, stable-2.14 is not used only by Ericsson but by many other clients that are not expecting any sort of regression or change in functionality on a stable branch.

If we allow active development, new functionalities and potential regressions on a stable branch, then why do we have branches at all?

I agree in making it work with Gerrit 2.14, 2.15 and master ... let's work towards that rather than jeopardizing a stable branch.

Patch-set: 1
Reviewer: Gerrit User 1006192 <1006192@173816e5-2b9a-37c3-8a2e-48639d4f1153>