Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> This will mean users on 2.14 cannot use the draft publishing event and velocity. (If they use the master branch)

With regards to velocity: yes, if you are using its-base on master it means you want to get the 'latest and greatest' status of its-* plugins you'll have to use Soy templates.
I'll need to add the Soy dependencies as well to make sure that can be built on 2.14.

Draft support is a different story: yes I need to add some support for that in this change on 2.14.

Patch-set: 3
Reviewer: Gerrit User 1006192 <1006192@173816e5-2b9a-37c3-8a2e-48639d4f1153>