Replace sets of Property by a simple map
So far, the information about an event was stored in a collection of
collections of properties. A Property was defined as a wrapper around a
simple pair of key -> value entries. Even if simple as concept, this
data structure made the code unnecessarily complicated. For example, in
order to get a value from the properties collection, it was needed to
iterate over the collection and compare the desired value with the
property key. When these properties were about to be used, as when
adding a Soy comment, i.e., this collection needed to be converted to a
map to be passed as the context to the Soy renderer. In the tests, lots
of mocking was done around the Property object which bloated and
complicated the test code.

Using a simple map for storing the event information fulfills the same
goal but with some advantages: simpler loops, no need to iterate over
a collection to find and get some values, no need to convert to a map
later when the info is needed and, besides that, simplified tests.

Change-Id: Id429d8cd9e626d88c1f81954e6b37488896f9b90