Update patch set 5
Patch Set 5:

> Patch Set 3:
> 
> > Patch Set 3: Code-Review-1
> > 
> > We should then do something in the init step as well: if there is still a deprecated file, we should fail the init and force the user to migrate.
> 
> In the init step only the project integration and branch configuration are set up, no mention to rules files. These are read only (unless I'm missing something) when the rules are initialized in the constructor of the RuleBase class.

I came up with this option: if a deprecated config file is present, the init step shows an error message asking the user to migrate the configuration and fails. What could be an issue is that now the init in its-base needs to know about the site path which means a new parameter should be passed and then all other plugins using the init should be modified (I'm OK with this but IIUC, you don't want to break backward compatibility even in master branch).

Patch-set: 5
Reviewer: Gerrit User 1019328 <1019328@173816e5-2b9a-37c3-8a2e-48639d4f1153>