Properly escape the subject
This is needed to properly escape the subject.

We need to do this to prevent Revert "test" failing.

When i use escapeJava in combination with soy's function escapeUri
this properly turned "test" into \"test\" and then escapeUri turned it back
into "test". This properly escapes as for example

Test+Test does not get converted into Test Test now.

This adds a new var escapedSubject to make sure this change is compat
for users workflow, ie no breaking change.

This makes it possible to backport to stable branches which is needed.

Bug: Issue 8432
Change-Id: Iecd34d826d5ee471dbd247078ab9a99d062cd091