Update patch set 9
Patch Set 9:

> Patch Set 9: Code-Review-2
> 
> Every change should do one useable thing and should give some value. If this is only focussed to support another change, why don't you squash them together or at least put them into the same topic?

Refactoring should be done separately from adding new functionality.  This is what is always requested in reviews on the gerrit project.

In this case the changes probably don't even need to be on the same topic; merging this one before the others it won't break anything.

Not necessary to put CR-2 IMHO.

Patch-set: 9
CC: Gerrit User 1011123 <1011123@173816e5-2b9a-37c3-8a2e-48639d4f1153>