Update patch set 9
Patch Set 9:

> Patch Set 9:
> 
> > Patch Set 9: Code-Review-2
> > 
> > Every change should do one useable thing and should give some value. If this is only focussed to support another change, why don't you squash them together or at least put them into the same topic?
> 
> Refactoring should be done separately from adding new functionality.  This is what is always requested in reviews on the gerrit project.

True, but they should be in the same topic if the refactoring doesn't bring any value and doesn't make sense if not related to the other Change.

It is different if the refactoring improves the code readability, structure, maintainability :-)

> 
> In this case the changes probably don't even need to be on the same topic; merging this one before the others it won't break anything.
> 
> Not necessary to put CR-2 IMHO.

We don't have any other way to express vetos at the moment. It is not a "I don't like this refactoring" but rather "It should be submitted in a different way".

Patch-set: 9
Reviewer: Gerrit User 1006192 <1006192@173816e5-2b9a-37c3-8a2e-48639d4f1153>