Update patch set 9
Patch Set 9:

> > We don't have any other way to express vetos at the moment. It is not a "I don't like this refactoring" but rather "It should be submitted in a different way".
> 
> That's what -1 is for.  IMHO -2 is for "this is completely wrong and will never be accepted"

mmm ... that's not what the description says :-)
"Do not submit" doesn't mean that is completely wrong. Just means that it should not be submitted as-is.

See for instance  I60139e99c where -2 is used as veto just to make sure the change is not merged until is "as perfect as possible", that doesn't mean that is completely wrong :-)

Patch-set: 9
Reviewer: Gerrit User 1006192 <1006192@173816e5-2b9a-37c3-8a2e-48639d4f1153>