Update patch set 9
Patch Set 9: Code-Review-2

(2 comments)

Even though *I really like* this refactoring as the code is a lot easier to follow, I need to put a -2 to express the veto on merging this change until we have verified that all the other its-* plugins are still working properly.

Modifying a public interface forces to revisit the impact on the other plugins, especially because its-base is NOT a plugin but only defines a hierarchy for all the other its-* integrations.

Patch-set: 9
Reviewer: Gerrit User 1006192 <1006192@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Code-Review=-2