Update patch set 9
Patch Set 9:

> Patch Set 9:
> 
> > > We don't have any other way to express vetos at the moment. It is not a "I don't like this refactoring" but rather "It should be submitted in a different way".
> > 
> > That's what -1 is for.  IMHO -2 is for "this is completely wrong and will never be accepted"
> 
> mmm ... that's not what the description says :-)
> "Do not submit" doesn't mean that is completely wrong. Just means that it should not be submitted as-is.

OK, let me rephrase.  It means "This cannot be submitted as is because it either doesn't work properly or will break something".

That is not the case for this change. It doesn't break anything; it's simply a refactoring that you are not in favour of.

> 
> See for instance  I60139e99c where -2 is used as veto just to make sure the change is not merged until is "as perfect as possible", that doesn't mean that is completely wrong :-)

In this case Hugo's concern is that it will destabilise the plugin.  To be fair, it's probably also not necessary to put +2 there either.

Patch-set: 9