Update patch set 9
Patch Set 9: Code-Review-1

> OK, let me rephrase.  It means "This cannot be submitted as is because it either doesn't work properly or will break something".
> 
> That is not the case for this change. It doesn't break anything; it's simply a refactoring that you are not in favour of.

Not exactly :-) I just said that the refactoring should be staying in a topic with the change is needs it and give it context to.

I wouldn't like to have this refactoring reviewed alone and merged as-is.

> > 
> > See for instance  I60139e99c where -2 is used as veto just to make sure the change is not merged until is "as perfect as possible", that doesn't mean that is completely wrong :-)
> 
> In this case Hugo's concern is that it will destabilise the plugin.  To be fair, it's probably also not necessary to put +2 there either.

-2 you mean ;-)

I believe he did just as placeholder to make sure that it doesn't get merged without having verified all the corner cases.

Anyway, happy to turn this into a -1 with a veto :-)

Patch-set: 9
Reviewer: Gerrit User 1006192 <1006192@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Code-Review=-1