Update patch set 8
Patch Set 8:

> Patch Set 8: Code-Review-1
> 
> How come that by changing 3 lines of a private method, which in theory should preserve the existing behaviour, causes so many changes in the tests?

The thing is, we were re-creating the link (formatChangeUrl) based on the changeUrl (AddStandardComment, line 65) but given the link is already present in the properties (it was calculated when extracting the properties from the event), we do not need to create it again. 

The tests were checking this creation by mocking the its.createLinkForWebui() method response but all of this code is no longer necessary as we can add directly the formatChangeUrl property to the message.

Patch-set: 8
Reviewer: Gerrit User 1019328 <1019328@173816e5-2b9a-37c3-8a2e-48639d4f1153>