Update patch set 20
Patch Set 20:

The implementation of multi-server on its-jira [1] is using the "default methods" version of this change (latest patch set), even if patch set 16 also explored the way you proposed (using a new interface). One  step we need in order to be able to advance in the multi-server implementation is to take the decision which is the preferred way of implementing this capacity in its-base. Could you please take a look again and let us know?

[1] https://gerrit-review.googlesource.com/c/plugins/its-jira/+/156092

Patch-set: 20
Reviewer: Gerrit User 1019328 <1019328@173816e5-2b9a-37c3-8a2e-48639d4f1153>